backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-genai==1.7.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.22.0
google-cloud-firestore==2.20.1
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
pydantic==2.10.6
